ex05

+++ modulo2/ex06/asm.h
@@ -1,4 +1,4 @@
 #ifndef ASM_H 
 #define ASM_H 
- int swapBytes(); 
+ short swapBytes(); 
 #endif
